requirements.txt
pytest==7.2.2
python-dateutil==2.8.2
